requirements.txt
fastapi
uvicorn[standard]
pydantic[dotenv]
passlib[bcrypt]
python-multipart
sqlalchemy
asyncpg
psycopg2-binary
alembic
minio
python-dotenv
pillow
cuid2
